Use a property for the cli engine config that doesn't conflict with the rc module's added properties. (#1)
The `rc` module adds the `configs` and `config` properties to the returned object to let the caller know where it actually found the configuration. It looks like this:

```
{
  "port": 9000,
  "mode": "test",
  "foo": "barbar",
  "something": "else",
  "_": [],
  "config": "config.json",
  "configs": [
    "/Users/stephen/repos/conftest/.myapprc",
    "config.json"
  ]
}
```

Sadly, this stomps on any key named `config` from the actual configuration rc is trying to load. So, eslint-output's config option was getting overwritten by the rc library.

This change makes eslint-output use a different key, so it doesn't get overwritten. Woop woop.